+++ HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 +++++++++
 - Output error message when downloading non-existing files in `ia download` rather
   than raising Python exception.
+- Fixed IOError in `ia search` when using `head`, `tail`, etc..
 
 0.7.7 2014-12-17
 ++++++++++++++++